backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
pydantic==2.9.0
sqlalchemy==2.0.35
alembic==1.13.2
supabase==2.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
aiofiles==24.1.0
python-dotenv==1.0.0
cryptography==43.0.0
geopy==2.4.1
opencage==3.0.2
